Add: Index class to load and use saved indices
The main information retrieval (IR) class. This class uses the
previously created index files to create an Index object. This object
represents the information retrieval (IR) portion of the system.
Individual page objects can be retrieved from disk, or a set of pages
that match a term list (either by union or intersection).

Also fixed a bug in the __str__ method of the Page class.

In addition, modified the Page.ID and Page.start fields to be of type
int instead of string.