Add: web user interface and Fix: regularize()
Still very simple and rough around the edges, but it works for now.

Currently shows the set of pages that represent the intersection of the
query terms. In addition, up to the first three sentences of the page
are displayed if the sentence contain at least one query term.

In addition, made part of the Page.regularize_text() method a public
function in indexer. In the process fixed a few bugs arising from using
the old value of token for each stage of the process.