Add: ranked list page retrieval using TF–IDF
Also updated the web UI to use the ranking information. The web UI also
now links to the simple wikipedia page online. Note: This presumes that
we are using the Simple English Wikipedia dump, so TODO is to make the
URL prefix configurable.

For more information on how to compute TF–IDF see:
http://en.wikipedia.org/wiki/Tf%E2%80%93idf

For more information about cosine similarity---and how to compare two
TF-IDF vectors---see:
https://secure.wikimedia.org/wikipedia/en/wiki/Vector_space_model

Updated the system components doc with the new TF-IDF functionality.