API: change return type of Index methods
The various page finding methods of Index, such as union(), intersect(),
and ranked(), used to return a list of Page objects. Sometimes it is not
necessary to retrieve the Page contents from disk. So, instead those
functions now return a list of Page.IDs, which can then be turned into
Page objects by using the get_page() method.

The ranked() method now returns a list of tuples in the form (Page.ID,
cosine_similarity).