CONF: web UI now listens on port 8080 by default
Changed port the web UI server listens on from 8888 to 8080 since we had
issues with being able to access the server externally on the old 8888
port.

Expanded the convenience listen() method provided by
tornado.web.Application to explicitly create the HTTPServer object. This
might be useful later if we need to modify some of the parameters of the
HTTPServer.