ADD: web UI now handles requests asynchronously
Before we handled each user's request all at once. However, this causes
the main Tornado IO loop to block meaning it cannot handle ANY other
requests!

By using the apply_async() method of multiprocessing pool, we are able
to hand off the CPU intensive portion of the work to a different process
and wait for the results. In the mean time, we can handle requests from
other users.

Future work would be to parallelize the CPU intensive portion of the
work, which would result in the average response time per request to
become shorter. Currently we spawn P processes, which means we can
handle up to P users and their queries at once. However, it might be
better to have less of these handler processes and more worker processes
that divide up the data.