Add: DocI class to access from disk and not RAM
Wrapper class around the .doci index file; allows for doci[ID].

This class allows the .doci {page.ID -> page.token_count} index to stay
on disk without having to load the entire file into memory. There is a
memory-speed tradeoff, but the speed lost with respect to the amount of
memory saved for larger corpora will be noticeable.

IMPORTANT: Save the resulting dictionary of a doci[Page.ID] call
locally! Because the dictionary has to be created for every
__getitem__() call, the operation becomes exponentially expensive!