Task: re-organize to make answer_engine module
Instead of having the web UI module handle the answer generation
process, we want to instead move that procedure to a separate module.

The answer engine module will be responsible for all the heavy lifting
after the corpus has been indexed and the user supplies the query. These
steps include:
* Question analysis
* Candidate document analysis
* Answer extraction
* Response generation

The web UI module will still be responsible for taking the answers from
the answer engine and formatting them for display on a web page.