Add: WordNet to match query senses with answer
The basic idea is to make sure each word in the query has at least one
match in the answer text. The problem is when there is not an exact
match. Therefore we use WordNet to find if there is a similar term in
the answer. A sentence term is similar if any of its senses is related
to the sense of the query term. Right now we use LCH to determine how
related query terms and sentence terms are.

In a future update (TODO), it would be nice to allow the user to
disambiguate their query terms so that instead of matching *any* of its
possible terms, a sentence term must match only a single sense.

Note: There may be a bug somewhere, because each AnswerEngine process
uses nearly as much RAM as the main process that holds the entire Index
in memory. I'm not sure why this is the case since the AnswerEngine
doesn't store the Index object after leaving __init__().