Task: store more query vars; more docstring info
Store more of the query arguments, such as num_top and start, in the
AnswerEngine object. They can be retrieved and used later, such as when
displaying the UI the user can be given a link to the next page of
results.

Also, earlier commit forgot to include docstring info about LCH.