API: use Gensim Dictionary for token mapping
Using the Gensim library class Dictionary, we convert tokens from
strings to a unique integer through a mapping {token -> token_id}. In
order to do this, we must make a first pass over the entire corpus to
extract the Dictionary.

The goal of doing this is to reduce memory requirements, as now all
references to tokens (in the Index) are by integers (size = 24 bits)
instead of strings (size = 33 + (1 * length) bits).

In addition, now that we know the vocabulary in advanced we can
filter_extremes() and remove tokens that occur too infrequently (say,
appear in less than 5 documents) or too frequently (such as in more than
50% of all documents). This helps reduce the footprint of the Index even
more.

Keep in mind that this is just for IR purposes, such as retrieving a
ranked list of documents, but also for other topic modeling algorithms
like LSA and LDA.