Add: plain-text corpus generation and paragraphs
Since we are taking a first pass anyway in order to build our
Dictionary, now we also save the preprocessed pages into a new
plain-text corpus file. This file then serves to be our new dataset, as
we retrieve our pages directly from here. The advantage is the file has
already had all of the MediaWiki markup removed, and the resulting file
is quite a bit smaller.

The format of the plain-text corpus has one page per line. The fields
are tab separated. In addition, the paragraphs in the page text are
labeled with the Unicode paragraph separator character (\u2029), and the
sentences are already segmented and identified by the Unicode line
separator character (\u2028).

The sentence segmentation process isn't that slow to begin with, so our
execution time savings are probably marginal, however going through the
process forced us to change how the Page class behaves. Instead of
containing sentences, the Page class now contains Paragraphs, which in
turn contain the sentences that make up that paragraph. Paragraphs can
be used in the future as part of a better candidate answer generator.

Some possible future considerations might include:
* Attempting to modularize the code that is shared between the first and
second pass functions that start the worker processes. This has been
tried in the past but ran into technical issues that I couldn't resolve.
* Right now we are making many copies of the Page text in the form of
the full original, full paragraph original, each sentence original, and
finally the tokenized copy of the words. Instead, we could save some
memory and perhaps some processing time not having to make so many
copies (though I haven't tried profiled this) if we kept only one
original and the other copies were stored by spans (index start-end).