Add: use MongoDB for the inverted index (toki)
So far we have been storing the (toki) inverted index {token ->
set(documents)} entirely in memory as a dictionary. This works fine for
small corpora, however this method breaks down if the dictionary cannot
fit in memory. The MongoDB provides a solution by storing an equivalent
to our dictionary on disk so that it doesn't need to be loaded into
memory.

During index creation, write performance is much slower than working
with a variable in memory. We should try adding threads to the writer
process to improve performance.

Read performance is also degraded at query time, however we only need to
send as many requests as there are tokens in the query. We also needed
to change the TF-IDF calculation to use the (Gensim) Dictionary instance
to retrieve the number of documents a word appears in, rather than the
old method of counting the size of the set of documents from the
in-memory toki index.