Add: thread for MongoDB writing
In an attempt to increase performance in writing to the MongoDB, we now
use a thread in the writer process to send updates.

This was tested with multiple threads, however this resulted in poor
performance over a single thread. Though, if the MongoDB is sharded then
multiple threads might improve performance.

Additionally, through testing it was determined sending roughly 8
updates at a time resulted in a 10% increase in performance over sending
one request per token. This number was tested from 8 - 128, but using
higher numbers didn't seem to improve too much.

It's worth repeating that whereas before the indexing process was
CPU-bound from processing the Pages, we are now I/O bound again because
of communication with the MongoDB. This cannot be avoided because we
can't fit the entire token-index into memory. However, we could try
sharding the MongoDB to achieve horizontal scaling, which might improve
our performance.