Fix: massive performance speedup in MongoDB
Turns out we were going about creating the inverted index (toki)
entirely the wrong way! We were trying to store items in the MongoDB the
same way we were creating the token-index (toki) for use in memory
{token -> set(Page.IDs)}. Instead, we just need to store something a lot
like the document-term matrix (doci) {Page.ID -> set(tokens)}, and then
let MongoDB create an index on the tokens internally.

This results in several orders of magnitude in speedup. Now we are back
to being CPU bound instead of I/O bound!