Fix: speedup in IR search for MongoDB
This is for the Union IR function. Previously we sent each query token
to the MongoDB and unioned the set of matching pages for each term. Now
we send all the query terms to the MongoDB in a single request. A page
matches if it contains any of the query terms sent.