Add: answer ranking and partial answer matching
Answers are now scored based on several different features. The value of
the features are added to the score by multiplying the value with a
weight. Currently the weights are chosen arbitrarily (best guess),
however using a machine learning (ML) algorithm to choose optimal
weights would be beneficial.

Because answers are ranked and no longer boolean, partial matching is
available.