Fix: ignore terms not in term-to-token dictionary
Previously, if a term not present in the Gensim dictionary---a mapping
from token string to an integer token ID---an error would occur. We now
ignore terms not in our token-dictionary when querying the IR system.

Also, use a copy of the term-argument so that the original is not
changed unexpectedly.