Add: machine learning to rerank answers
A corpus of questions (currently 19) with the correct answers
(question-answer-pairs.txt) was created in order to evaluate the system.
All answers with at least one matching term in the text was written to a
log file (log_training.txt), and the rank(s) of the correct answer(s)
was manually labeled. All other answers are assumed to be incorrect.
Mean reciprocal rank (MRR) can be used as a performance metric.

Previously, each feature in the scoring process was assigned an
arbitrary weight (based on research papers in the QA field). However, it
would seem worthwile to use machine learning in an attempt to optimize
the feature weights, hopefully resulting in improved answer rankings.
This is sometimes called learning to rerank in this research area.
Therefore, we wrote a script (log_training_to_arff.py) to convert the
question-answer pairs log file to a Weka format ARFF file in order to
train a classifier and learn weights for the features. We currently use
logistic regression, though SVM might be worthwhile as well.

In order to get useful results, we had to oversample the minority class
(correct answers) because the class distribution is so imbalanced that
there are vastly more incorrect examples. Without doing this, the
classifier simply labels everything as an incorrect answer, which is not
helpful. We tried undersampling the majority class (incorrect answers),
however the results were still not as useful.