Task: change defaults for Simple English Wiki
Since we are heavily using the Simple English version of Wikipedia, we
don't really need to use PyMongo, so we have changed the default to not
use it (still gives a warning if the library is not installed).

In addition, we don't need to write a log for training a machine
learning classifier by default---only when we are actively trying to
improve the features or weights---therefore we disable this by default
(can be enabled per query through the "train" GET argument).