Add: Attribution-NonCommercial-ShareAlike license
Basically, you can use the code of this project as long as you give me
credit for my work (attribution), and that you do not use this work for
any commercial purposes. In addition, if you choose to share this or
derived work, you may distribute the resulting work only under the same
or similar license to this one.

Please read the details of the license here:
http://creativecommons.org/licenses/by-nc-sa/3.0/