Fix: Wiktionary link conversion in Wiki markup
We use the WikiExtractor.py library in order to filter out the MediaWiki
Wiki markup from the articles. Unfortunately, the library simply removes
any word that has a link to the Wiktionary page (so that people can
quickly lookup a translation in their language). This caused many
sentences to have missing words, often keywords.

To correct this, in the remove_markup() function we have added a regular
expression to convert both the regular Wiktionary links, as well as the
"broken" Wiktionary links to just the word they are referencing. After
doing this conversion, we then run the article through the
WikiExtractor.py library to finish filtering out the markup (the
conversion to plain-text).