Task: no longer filter dictionary by default
Previously, the dictionary generated from the corpus defaulted to
removing terms that were either too rare or too common. Unfortunately,
the default settings caused too many terms to be removed, causing more
than a few questions to be unanswerable because the IR engine couldn't
find the correct page (if any at all) for the query terms.

Currently, we have disabled filtering of the dictionary. On small
corpora, like the Simple English Wikipedia, the dictionary file isn't
very large to begin with even if all terms are kept.

A future update is planned to make the dictionary filtering options an
option in the configuration file. In anticipation, the commented out
line now contains better default settings for the filtering step (if
enabled).